Update RobotFeedback.py
In fw 7.0.6: Upon Connect Joint and Cartesian Pose are sent. We can use getData() without parsing the received response as getData() has its own receive response logic to manage broken packets
In 8.x: Everytime upon connect, RobotStatus and GripperStatus are published. So we are receiving response and parsing it to _getRobotStatus and _getGripperStatus. getData doesn't check for Robot and Gripper status for every getData.